[chore](shared/TokenCom): Migrate TokenCom component
