+++ packages/sqle/src/icon/Scan.tsx
@@ -14,20 +14,3 @@ export const IcoScanId = () => {
     </svg>
   );
 };
-
-export const IconToken = () => {
-  return (
-    <svg
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M11.3333 9.33333H8.43906C7.88992 10.8869 6.40828 12 4.66666 12C2.45752 12 0.666656 10.2091 0.666656 8C0.666656 5.79086 2.45752 4 4.66666 4C6.40828 4 7.88992 5.11308 8.43906 6.66667H15.3333V9.33333H14V12H11.3333V9.33333ZM4.66666 9.33333C5.40304 9.33333 5.99999 8.7364 5.99999 8C5.99999 7.2636 5.40304 6.66667 4.66666 6.66667C3.93028 6.66667 3.33332 7.2636 3.33332 8C3.33332 8.7364 3.93028 9.33333 4.66666 9.33333Z"
-        fill="#8A8F99"
-      />
-    </svg>
-  );
-};
